% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_scripts.R
\name{convert_chrom}
\alias{convert_chrom}
\title{Converts chromatograms to long format}
\usage{
convert_chrom(x)
}
\arguments{
\item{x}{A chromatographic matrix in wide format.}
}
\value{
A chromatographic matrix in long format.
}
\description{
Converts chromatograms to long format
}
\author{
Ethan Bass
}
