% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_scripts.R
\name{summarize_peak_info}
\alias{summarize_peak_info}
\title{Summarizes peak information}
\usage{
summarize_peak_info(peak_table)
}
\arguments{
\item{peak_table}{A \code{peak_table} object.}
}
\value{
A \code{peak_table} object.
}
\description{
Summarizes peak information
}
\author{
Ethan Bass
}
