% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessory_scripts.R
\name{tidy_chrom_converter}
\alias{tidy_chrom_converter}
\title{Converts list of chromatograms to long format}
\usage{
tidy_chrom_converter(x)
}
\arguments{
\item{x}{A list of chromatographic matrices in wide format.}
}
\value{
A list of chromatographic matrices in long format.
}
\description{
Converts list of chromatograms to long format
}
\author{
Ethan Bass
}
