% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromViewer.R
\name{chrom_viewer}
\alias{chrom_viewer}
\title{Shiny gadget for viewing chromatograms}
\usage{
chrom_viewer(peak_table, chrom_list)
}
\arguments{
\item{peak_table}{A \code{peak_table} object.}

\item{chrom_list}{A list of chromatograms.}
}
\value{
No return value.
}
\description{
Shiny gadget for viewing chromatograms
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(Sa_warp)
data(pk_tab)
chrom_viewer(pk_tab)
\dontshow{\}) # examplesIf}
}
\author{
Ethan Bass
}
